requirements.txt
alembic==0.9.1
appdirs==1.4.3
chardet==2.3.0
dataset==0.8.0
diff-match-patch==20121119
dj-database-url==0.4.2
Django==1.10.6
django-debug-toolbar==1.7
django-debug-toolbar-request-history==0.0.4
django-import-export==0.5.1
django-mongoengine-forms==0.4.1
django-widget-tweaks==1.4.1
djangorestframework==3.3.3
Mako==1.0.6
MarkupSafe==1.0
mongoengine==0.10.6
normality==0.4.0
olefile==0.44
packaging==16.8
Pillow==4.0.0
psutil==4.3.0
psycopg2==2.7.1
pymongo==3.2.2
pyparsing==2.2.0
PyPDF2==1.26.0
pypng==0.0.18
python-dateutil==2.5.3
python-decouple==3.0
python-editor==1.0.3
pytz==2016.4
PyYAML==3.12
six==1.10.0
SQLAlchemy==1.1.7
sqlparse==0.2.3
tablib==0.11.4
Unipath==1.1
virtualenv==15.0.2
